Executive decision: /usr/local/libexec.
This is "the right place" on RHEL/CentOS, but the directory doesn't
necessarily exist on other distros. It would be nice for this to be
configurable, but I want as few moving parts in the privop executable
as possible. We could override this at compile time, e.g:

    https://github.com/Ropes/go-linker-vars-example

but I'll leave that as a TODO